chore: cleanup unused dependencies in internal + moved route logic to separate file
